refactor(hpm): expose CAN driver as HPMCAN
