+++ driver/hpm/hpm_can.hpp
@@ -9,28 +9,31 @@
 #include "hpm_soc.h"
 #include "queue.hpp"
 
-#if __has_include("hpm_can_drv.h") && __has_include("hpm_can_regs.h")
-#include "hpm_can_drv.h"
+#if __has_include("hpm_mcan_drv.h") && __has_include("hpm_mcan_regs.h") && \
+                                                     __has_include("hpm_mcan_soc.h")
+#include "hpm_mcan_drv.h"
 using LibXRHpmCanStatus = hpm_stat_t;
-using LibXRHpmCanMode = can_node_mode_t;
-using LibXRHpmCanTxMessage = can_transmit_buf_t;
-using LibXRHpmCanRxMessage = can_receive_buf_t;
-#if defined(CAN_SOC_MAX_COUNT) && (CAN_SOC_MAX_COUNT > 0) && defined(HPM_CAN0)
+using LibXRHpmCanMode = mcan_node_mode_t;
+using LibXRHpmCanTxFrame = mcan_tx_frame_t;
+using LibXRHpmCanRxFrame = mcan_rx_message_t;
+#if ((defined(MCAN_SOC_MAX_COUNT) && (MCAN_SOC_MAX_COUNT > 0)) || \
+     (defined(CAN_SOC_MAX_COUNT) && (CAN_SOC_MAX_COUNT > 0))) &&  \
+    defined(HPM_MCAN0)
 #define LIBXR_HPM_CAN_SUPPORTED 1
 #else
 #define LIBXR_HPM_CAN_SUPPORTED 0
 #endif
 #else
 #define LIBXR_HPM_CAN_SUPPORTED 0
-using CAN_Type = void;
+using MCAN_Type = void;
 using LibXRHpmCanStatus = int;
 enum LibXRHpmCanMode : int
 {
 };
-struct LibXRHpmCanTxMessage
+struct LibXRHpmCanTxFrame
 {
 };
-struct LibXRHpmCanRxMessage
+struct LibXRHpmCanRxFrame
 {
 };
 #endif
@@ -40,10 +43,10 @@ namespace LibXR
 
 /**
  * @class HPMCAN
- * @brief HPM classic CAN driver implementation.
+ * @brief HPM CAN driver implementation for classic LibXR CAN frames.
  *
- * Wraps HPM SDK's `CAN_Type` controller as the LibXR classic CAN interface.
- * Board pin mux and transceiver standby pins are intentionally left to board code.
+ * This class intentionally exposes the classic `CAN` interface first. CAN FD can be
+ * layered on top of the same HPM SDK primitives later through LibXR::FDCAN.
  */
 class HPMCAN final : public CAN
 {
@@ -51,13 +54,21 @@ class HPMCAN final : public CAN
   static constexpr uint32_t INVALID_IRQ = 0xFFFFFFFFu;
 
   /**
-   * @brief Construct one HPM classic CAN driver.
+   * @brief Construct one HPM CAN driver.
    * @param can CAN peripheral base.
    * @param clock CAN peripheral clock name.
    * @param irq IRQ number for this CAN instance. Pass `INVALID_IRQ` to skip PLIC setup.
    * @param queue_size TX software queue size.
    */
-  HPMCAN(CAN_Type* can, clock_name_t clock, uint32_t irq, uint32_t queue_size);
+  HPMCAN(MCAN_Type* can, clock_name_t clock, uint32_t irq, uint32_t queue_size);
+
+  /**
+   * @brief Construct one HPM CAN driver with message RAM.
+   * @param msg_buf Message RAM buffer base. HPM5E31 requires this in `.ahb_sram`.
+   * @param msg_buf_size Message RAM buffer size in bytes.
+   */
+  HPMCAN(MCAN_Type* can, clock_name_t clock, uint32_t irq, uint32_t queue_size,
+          void* msg_buf, uint32_t msg_buf_size);
   ~HPMCAN() override = default;
 
   ErrorCode SetConfig(const CAN::Configuration& cfg) override;
@@ -65,7 +76,7 @@ class HPMCAN final : public CAN
   ErrorCode AddMessage(const ClassicPack& pack) override;
   ErrorCode GetErrorState(CAN::ErrorState& state) const override;
 
-  /** @brief Drain RX, service TX, and publish error frames for this CAN IRQ. */
+  /** @brief Drain RX FIFOs, service TX, and publish error frames for this CAN IRQ. */
   void ProcessInterrupt();
 
   /** @brief Enable the CAN IRQ in the HPM interrupt controller. */
@@ -74,23 +85,34 @@ class HPMCAN final : public CAN
   /** @brief Disable the CAN IRQ in the HPM interrupt controller. */
   ErrorCode DisableInterrupt();
 
+  /**
+   * @brief Set message RAM before configuration.
+   *
+   * HPM5E31/HPM5E00 place CAN message RAM in AHB SRAM. Board code should pass
+   * a buffer placed in `.ahb_sram`, usually `MCAN_MSG_BUF_SIZE_IN_WORDS` words.
+   */
+  ErrorCode SetMessageBuffer(void* msg_buf, uint32_t msg_buf_size);
+
  private:
   static ErrorCode ConvertStatus(LibXRHpmCanStatus status);
   static LibXRHpmCanMode ConvertMode(const CAN::Mode& mode);
-  static void BuildTxMessage(const ClassicPack& pack, LibXRHpmCanTxMessage& message);
-  static void BuildRxPack(const LibXRHpmCanRxMessage& message, ClassicPack& pack);
-  static ErrorID ConvertErrorKind(uint8_t kind);
+  static void BuildTxFrame(const ClassicPack& pack, LibXRHpmCanTxFrame& frame);
+  static void BuildRxPack(const LibXRHpmCanRxFrame& frame, ClassicPack& pack);
+  static ErrorID ConvertLastError(uint8_t last_error);
 
   void EnableCanInterrupts();
   void DisableCanInterrupts();
+  ErrorCode ApplyMessageBuffer();
   void TxService();
-  void ProcessRx();
+  void ProcessRxFifo(uint32_t fifo_index);
   void ProcessTx();
-  void ProcessError(uint8_t error_flags, uint8_t last_error_kind);
+  void ProcessError();
 
-  CAN_Type* can_;
+  MCAN_Type* can_;
   clock_name_t clock_;
   uint32_t irq_;
+  void* msg_buf_{nullptr};
+  uint32_t msg_buf_size_{0};
 
   MPMCQueue<ClassicPack> tx_queue_;
   bool tx_retry_valid_{false};
